requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.1
frozenlist==1.6.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.23.0
Pygments==2.19.1
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
ta==0.11.0
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.0.2
yarl==1.20.0
